Changes the build script to generate release pdbs as requested from some devs.
* [qtbase]: Fixes a crash that occurred when a submenu was about to be shown and
  threw an exception within its construction, what showed up an assert dialog,
  what let Qt to internally close the menu - setting the d->activeMenu to a
  nullptr, that then crashed inside the setSubMenuPopup call.

* [qtimageformats]: merge of the libtiff 4.0.7 changes

Used to build the 3dsmax artifact Qt-5.6.2-3dsmax-006.7z
(build against windows 10.0.14393.0 runtime)

+++ adsk_3dsmax_build.py
@@ -152,7 +152,8 @@ if MODE == 'release' or MODE == 'debug_and_release':
                     # in a separate step
         # ' -icu ' +
         # ' -I \"' + ICU_INCLUDE_PATH + '\" -L \"' + ICU_LIB_PATH + '\"'
-        ' -make tests'
+        ' -make tests '
+        ' -force-debug-info '
         ' -release ', env=BUILD_ENV, shell=True)
 
     subprocess.check_call('jom', env=BUILD_ENV, shell=True)
@@ -166,8 +167,9 @@ if MODE == 'release' or MODE == 'debug_and_release':
 # -----------------------------------------------------------------------------
 for f in os.listdir(os.path.join(PREFIX_PATH, 'bin')):
     if f.endswith('.pdb') and not f.endswith('d.pdb'):
-        print 'deleting unwanted ' + f + '\n'
-        os.remove(os.path.join(PREFIX_PATH, 'bin', f))
+        if os.path.exists( os.path.join(PREFIX_PATH, 'bin', f[:-4] + '.exe' ) ):
+            print 'deleting unwanted ' + f
+            os.remove(os.path.join(PREFIX_PATH, 'bin', f))
 # -----------------------------------------------------------------------------
